Update abstract_detection_method.py
Missed the plus 1 in the previous edit. This change reflects the need to capture active emissions throughout the current time-step. These emissions may begin prior to the current time-step and end in a future time-step, begin before the time-step and end in the current time-step, or begin in the current time-step and end in a future time-step.